Adds eager evaluation based on expression traits criteria
Those 'criteria' are a bit rudimnetary but most of the pieces are in
place now to make more reasoned lazy-vs-eager decisions.

The placeholder criteria with this commit is to eagerly evaluate a
sub-Expression if the parent Expression it is nested within would access
its coefficients more than once, on average, while evaluating that
parent Expression. For the current, limited set of LA operations, this
effectively means that the operands to matrix multiplications will be
eagerly evaluated.

As I type this, it comes to mind that this implementation will eagerly
evaluate the first sub-Expressions that meet these criteria, but
sub-Expressions of those will be lazily evaluated during that eager
evaluation.

Translation: If we have nested matrix multiplications, the top-most
matrix multiplication operands will be eagerly evaluated. But in
performing that eager evaluation, the bottom matrix multiplication
operands will be lazily evaluated.

Better is to eagerly evaluate everything that makes sense from the
bottom up. How to achieve?

While heading down the Expression looking for the next operands
to eagerly evaluate, when one is found, continue to search below it BUT
reset the criteria AS IF the first to-be eagerly evaluated operand is at
the top of the Expression.